@@ -1,27 +1,27 @@
 {
-  "_from": "mime-types@~2.1.18",
-  "_id": "mime-types@2.1.22",
+  "_from": "mime-types@~2.1.24",
+  "_id": "mime-types@2.1.25",
   "_inBundle": false,
-  "_integrity": "sha512-aGl6TZGnhm/li6F7yx82bJiBZwgiEa4Hf6CNr8YO+r5UHr53tSTYZb102zyU50DOWWKeOv0uQLRL0/9EiKWCog==",
+  "_integrity": "sha512-5KhStqB5xpTAeGqKBAMgwaYMnQik7teQN4IAzC7npDv6kzeU6prfkR67bc87J1kWMPGkoaZSq1npmexMgkmEVg==",
   "_location": "/accepts/mime-types",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "mime-types@~2.1.18",
+    "raw": "mime-types@~2.1.24",
     "name": "mime-types",
     "escapedName": "mime-types",
-    "rawSpec": "~2.1.18",
+    "rawSpec": "~2.1.24",
     "saveSpec": null,
-    "fetchSpec": "~2.1.18"
+    "fetchSpec": "~2.1.24"
   },
   "_requiredBy": [
     "/accepts"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.22.tgz",
-  "_shasum": "fe6b355a190926ab7698c9a0556a11199b2199bd",
-  "_spec": "mime-types@~2.1.18",
-  "_where": "K:\\שרגא\\פרויקט סיום יד\\ProjectExample\\Client\\node_modules\\accepts",
+  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.25.tgz",
+  "_shasum": "39772d46621f93e2a80a856c53b86a62156a6437",
+  "_spec": "mime-types@~2.1.24",
+  "_where": "C:\\Users\\1\\Desktop\\planner4u\\Planner4U\\Client\\node_modules\\accepts",
   "bugs": {
     "url": "https://github.com/jshttp/mime-types/issues"
   },
@@ -43,19 +43,19 @@
     }
   ],
   "dependencies": {
-    "mime-db": "~1.38.0"
+    "mime-db": "1.42.0"
   },
   "deprecated": false,
   "description": "The ultimate javascript content-type utility.",
   "devDependencies": {
-    "eslint": "5.13.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-node": "7.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-standard": "4.0.0",
-    "mocha": "5.2.0",
-    "nyc": "13.3.0"
+    "eslint": "6.6.0",
+    "eslint-config-standard": "14.1.0",
+    "eslint-plugin-import": "2.18.2",
+    "eslint-plugin-node": "10.0.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.0.1",
+    "mocha": "6.2.2",
+    "nyc": "14.1.1"
   },
   "engines": {
     "node": ">= 0.6"
@@ -82,5 +82,5 @@
     "test-cov": "nyc --reporter=html --reporter=text npm test",
     "test-travis": "nyc --reporter=text npm test"
   },
-  "version": "2.1.22"
+  "version": "2.1.25"
 }
