@@ -1,8 +1,8 @@
 {
   "_from": "accepts@~1.3.5",
-  "_id": "accepts@1.3.5",
+  "_id": "accepts@1.3.7",
   "_inBundle": false,
-  "_integrity": "sha1-63d99gEXI6OxTopywIBcjoZ0a9I=",
+  "_integrity": "sha512-Il80Qs2WjYlJIBNzNkK6KYqlVMTbZLXgHx2oT0pU/fjRHyEp+PEfEPY0R3WCwAGVOtauxh1hOxNgIf5bv7dQpA==",
   "_location": "/accepts",
   "_phantomChildren": {},
   "_requested": {
@@ -20,10 +20,10 @@
     "/express",
     "/serve-index"
   ],
-  "_resolved": "https://registry.npmjs.org/accepts/-/accepts-1.3.5.tgz",
-  "_shasum": "eb777df6011723a3b14e8a72c0805c8e86746bd2",
+  "_resolved": "https://registry.npmjs.org/accepts/-/accepts-1.3.7.tgz",
+  "_shasum": "531bc726517a3b2b41f850021c6cc15eaab507cd",
   "_spec": "accepts@~1.3.5",
-  "_where": "K:\\שרגא\\פרויקט סיום יד\\ProjectExample\\Client\\node_modules\\compression",
+  "_where": "C:\\Users\\1\\Desktop\\planner4u\\Planner4U\\Client\\node_modules\\compression",
   "bugs": {
     "url": "https://github.com/jshttp/accepts/issues"
   },
@@ -40,21 +40,22 @@
     }
   ],
   "dependencies": {
-    "mime-types": "~2.1.18",
-    "negotiator": "0.6.1"
+    "mime-types": "~2.1.24",
+    "negotiator": "0.6.2"
   },
   "deprecated": false,
   "description": "Higher-level content negotiation",
   "devDependencies": {
-    "eslint": "4.18.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.9.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.6.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "~1.21.5"
+    "deep-equal": "1.0.1",
+    "eslint": "5.16.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.17.2",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "8.0.1",
+    "eslint-plugin-promise": "4.1.1",
+    "eslint-plugin-standard": "4.0.0",
+    "mocha": "6.1.4",
+    "nyc": "14.0.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -80,8 +81,8 @@
   "scripts": {
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --check-leaks --bail test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "test-travis": "nyc --reporter=text npm test"
   },
-  "version": "1.3.5"
+  "version": "1.3.7"
 }
