@@ -1,8 +1,8 @@
 {
   "_from": "ajv-keywords@^3.1.0",
-  "_id": "ajv-keywords@3.4.0",
+  "_id": "ajv-keywords@3.4.1",
   "_inBundle": false,
-  "_integrity": "sha512-aUjdRFISbuFOl0EIZc+9e4FfZp0bDZgAdOOf30bJmw8VM9v84SHyVyxDfbWxpGYbdZD/9XoKxfHVNmxPkhwyGw==",
+  "_integrity": "sha512-RO1ibKvd27e6FEShVFfPALuHI3WjSVNeK5FIsmme/LYRNxjKuNj+Dt7bucLa6NdSv3JcVTyMlm9kGR84z1XpaQ==",
   "_location": "/ajv-keywords",
   "_phantomChildren": {},
   "_requested": {
@@ -20,10 +20,10 @@
     "/webpack",
     "/webpack/schema-utils"
   ],
-  "_resolved": "https://registry.npmjs.org/ajv-keywords/-/ajv-keywords-3.4.0.tgz",
-  "_shasum": "4b831e7b531415a7cc518cd404e73f6193c6349d",
+  "_resolved": "https://registry.npmjs.org/ajv-keywords/-/ajv-keywords-3.4.1.tgz",
+  "_shasum": "ef916e271c64ac12171fd8384eaae6b2345854da",
   "_spec": "ajv-keywords@^3.1.0",
-  "_where": "K:\\שרגא\\פרויקט סיום יד\\ProjectExample\\Client\\node_modules\\schema-utils",
+  "_where": "C:\\Users\\1\\Desktop\\planner4u\\Planner4U\\Client\\node_modules\\schema-utils",
   "author": {
     "name": "Evgeny Poberezkin"
   },
@@ -39,12 +39,12 @@
     "chai": "^4.2.0",
     "coveralls": "^3.0.2",
     "dot": "^1.1.1",
-    "eslint": "^5.0.0",
+    "eslint": "^6.0.0",
     "glob": "^7.1.3",
     "istanbul": "^0.4.3",
     "js-beautify": "^1.8.9",
     "json-schema-test": "^2.0.0",
-    "mocha": "^5.2.0",
+    "mocha": "^6.0.0",
     "pre-commit": "^1.1.3",
     "uuid": "^3.3.2"
   },
@@ -76,5 +76,5 @@
     "test-cov": "istanbul cover -x 'spec/**' node_modules/mocha/bin/_mocha -- spec/*.spec.js -R spec",
     "test-spec": "mocha spec/*.spec.js -R spec"
   },
-  "version": "3.4.0"
+  "version": "3.4.1"
 }
