@@ -1,27 +1,27 @@
 {
-  "_from": "mime-db@>= 1.38.0 < 2",
-  "_id": "mime-db@1.38.0",
+  "_from": "mime-db@>= 1.40.0 < 2",
+  "_id": "mime-db@1.42.0",
   "_inBundle": false,
-  "_integrity": "sha512-bqVioMFFzc2awcdJZIzR3HjZFX20QhilVS7hytkKrv7xFAn8bM1gzc/FOX2awLISvWe0PV8ptFKcon+wZ5qYkg==",
+  "_integrity": "sha512-UbfJCR4UAVRNgMpfImz05smAXK7+c+ZntjaA26ANtkXLlOe947Aag5zdIcKQULAiF9Cq4WxBi9jUs5zkA84bYQ==",
   "_location": "/compressible/mime-db",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "mime-db@>= 1.38.0 < 2",
+    "raw": "mime-db@>= 1.40.0 < 2",
     "name": "mime-db",
     "escapedName": "mime-db",
-    "rawSpec": ">= 1.38.0 < 2",
+    "rawSpec": ">= 1.40.0 < 2",
     "saveSpec": null,
-    "fetchSpec": ">= 1.38.0 < 2"
+    "fetchSpec": ">= 1.40.0 < 2"
   },
   "_requiredBy": [
     "/compressible"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.38.0.tgz",
-  "_shasum": "1a2aab16da9eb167b49c6e4df2d9c68d63d8e2ad",
-  "_spec": "mime-db@>= 1.38.0 < 2",
-  "_where": "K:\\שרגא\\פרויקט סיום יד\\ProjectExample\\Client\\node_modules\\compressible",
+  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.42.0.tgz",
+  "_shasum": "3e252907b4c7adb906597b4b65636272cf9e7bac",
+  "_spec": "mime-db@>= 1.40.0 < 2",
+  "_where": "C:\\Users\\1\\Desktop\\planner4u\\Planner4U\\Client\\node_modules\\compressible",
   "bugs": {
     "url": "https://github.com/jshttp/mime-db/issues"
   },
@@ -45,20 +45,20 @@
   "deprecated": false,
   "description": "Media Type Database",
   "devDependencies": {
-    "bluebird": "3.5.3",
+    "bluebird": "3.5.5",
     "co": "4.6.0",
     "cogent": "1.0.1",
-    "csv-parse": "3.2.0",
-    "eslint": "5.13.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-node": "7.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-standard": "4.0.0",
+    "csv-parse": "4.4.6",
+    "eslint": "6.4.0",
+    "eslint-config-standard": "14.1.0",
+    "eslint-plugin-import": "2.18.2",
+    "eslint-plugin-node": "10.0.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.0.1",
     "gnode": "0.1.2",
-    "mocha": "5.2.0",
-    "nyc": "13.2.0",
-    "raw-body": "2.3.3",
+    "mocha": "6.2.0",
+    "nyc": "14.1.1",
+    "raw-body": "2.4.1",
     "stream-to-array": "2.3.0"
   },
   "engines": {
@@ -97,5 +97,5 @@
     "update": "npm run fetch && npm run build",
     "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "1.38.0"
+  "version": "1.42.0"
 }
