@@ -1,46 +1,51 @@
 {
-  "_from": "content-disposition@0.5.2",
-  "_id": "content-disposition@0.5.2",
+  "_from": "content-disposition@0.5.3",
+  "_id": "content-disposition@0.5.3",
   "_inBundle": false,
-  "_integrity": "sha1-DPaLud318r55YcOoUXjLhdunjLQ=",
+  "_integrity": "sha512-ExO0774ikEObIAEV9kDo50o+79VCUdEB6n6lzKgGwupcVeRlhrj3qGAfwq8G6uBJjkqLrhT0qEYFcWng8z1z0g==",
   "_location": "/content-disposition",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "content-disposition@0.5.2",
+    "raw": "content-disposition@0.5.3",
     "name": "content-disposition",
     "escapedName": "content-disposition",
-    "rawSpec": "0.5.2",
+    "rawSpec": "0.5.3",
     "saveSpec": null,
-    "fetchSpec": "0.5.2"
+    "fetchSpec": "0.5.3"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/content-disposition/-/content-disposition-0.5.2.tgz",
-  "_shasum": "0cf68bb9ddf5f2be7961c3a85178cb85dba78cb4",
-  "_spec": "content-disposition@0.5.2",
-  "_where": "K:\\שרגא\\פרויקט סיום יד\\ProjectExample\\Client\\node_modules\\express",
+  "_resolved": "https://registry.npmjs.org/content-disposition/-/content-disposition-0.5.3.tgz",
+  "_shasum": "e130caf7e7279087c5616c2007d0485698984fbd",
+  "_spec": "content-disposition@0.5.3",
+  "_where": "C:\\Users\\1\\Desktop\\planner4u\\Planner4U\\Client\\node_modules\\express",
+  "author": {
+    "name": "Douglas Christopher Wilson",
+    "email": "doug@somethingdoug.com"
+  },
   "bugs": {
     "url": "https://github.com/jshttp/content-disposition/issues"
   },
   "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
-  ],
+  "dependencies": {
+    "safe-buffer": "5.1.2"
+  },
   "deprecated": false,
   "description": "Create and parse Content-Disposition header",
   "devDependencies": {
-    "eslint": "3.11.1",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.3.0",
-    "eslint-plugin-standard": "2.0.1",
+    "deep-equal": "1.0.1",
+    "eslint": "5.10.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-markdown": "1.0.0-rc.1",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "1.21.5"
+    "mocha": "5.2.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -65,10 +70,10 @@
     "url": "git+https://github.com/jshttp/content-disposition.git"
   },
   "scripts": {
-    "lint": "eslint .",
+    "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
   },
-  "version": "0.5.2"
+  "version": "0.5.3"
 }
