@@ -1,28 +1,28 @@
 {
-  "_from": "http-errors@~1.6.3",
-  "_id": "http-errors@1.6.3",
+  "_from": "http-errors@1.7.2",
+  "_id": "http-errors@1.7.2",
   "_inBundle": false,
-  "_integrity": "sha1-i1VoC7S+KDoLW/TqLjhYC+HZMg0=",
+  "_integrity": "sha512-uUQBt3H/cSIVfch6i1EuPNy/YsRSOUBXTVfZ+yR7Zjez3qjBz6i9+i4zjNaoqcoFVI4lQJ5plg63TvGfRSDCRg==",
   "_location": "/express/http-errors",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "http-errors@~1.6.3",
+    "raw": "http-errors@1.7.2",
     "name": "http-errors",
     "escapedName": "http-errors",
-    "rawSpec": "~1.6.3",
+    "rawSpec": "1.7.2",
     "saveSpec": null,
-    "fetchSpec": "~1.6.3"
+    "fetchSpec": "1.7.2"
   },
   "_requiredBy": [
     "/express/body-parser",
     "/express/raw-body"
   ],
-  "_resolved": "https://registry.npmjs.org/http-errors/-/http-errors-1.6.3.tgz",
-  "_shasum": "8b55680bb4be283a0b5bf4ea2e38580be1d9320d",
-  "_spec": "http-errors@~1.6.3",
-  "_where": "K:\\שרגא\\פרויקט סיום יד\\ProjectExample\\Client\\node_modules\\express\\node_modules\\body-parser",
+  "_resolved": "https://registry.npmjs.org/http-errors/-/http-errors-1.7.2.tgz",
+  "_shasum": "4f5029cf13239f31036e5b2e55292bcfbcc85c8f",
+  "_spec": "http-errors@1.7.2",
+  "_where": "C:\\Users\\1\\Desktop\\planner4u\\Planner4U\\Client\\node_modules\\express\\node_modules\\body-parser",
   "author": {
     "name": "Jonathan Ong",
     "email": "me@jongleberry.com",
@@ -45,21 +45,22 @@
   "dependencies": {
     "depd": "~1.1.2",
     "inherits": "2.0.3",
-    "setprototypeof": "1.1.0",
-    "statuses": ">= 1.4.0 < 2"
+    "setprototypeof": "1.1.1",
+    "statuses": ">= 1.5.0 < 2",
+    "toidentifier": "1.0.0"
   },
   "deprecated": false,
   "description": "Create HTTP error objects",
   "devDependencies": {
-    "eslint": "4.18.1",
-    "eslint-config-standard": "11.0.0",
-    "eslint-plugin-import": "2.9.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.6.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint": "5.13.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.16.0",
+    "eslint-plugin-markdown": "1.0.0",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "1.21.5"
+    "mocha": "5.2.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -82,10 +83,10 @@
     "url": "git+https://github.com/jshttp/http-errors.git"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint --plugin markdown --ext js,md . && node ./scripts/lint-readme-list.js",
     "test": "mocha --reporter spec --bail",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter dot"
   },
-  "version": "1.6.3"
+  "version": "1.7.2"
 }
